Add files via upload
This QuantConnect script takes Bitcoin, the S&P500 as a risky portfolio and bonds and gold as a riskless one. It looks at the economic environment using factors to see if the risky portfolio or the riskless should be invested in. It allocates weights to the portfolios using the Random Forest Regressor.